backend/requirements.txt
Django~=2.2.4
authentication
serializer
djangorestframework
djangorestframework-jwt
django-cors-headers
